# Cache str.len()