Rust str is utf-8, so it doesn't match to CPython str very well.
Probably we would reimplement PyString as same as CPython, but this is a
temporal fix for easy speedup.

Before:
    $ time cargo run --release Lib/test/test_unicode.py
    13.56s user 0.16s system 88% cpu 15.487 total

After:
    $ time cargo run --release Lib/test/test_unicode.py
    10.02s user 0.10s system 88% cpu 11.448 total