@coolreader18 2020-04-29
Is there anywhere that uses `.chars().count()` of a PyString to do some operation that we could replace with a call to `len()`?