@youknowone 2020-04-29
Unfortunately, it is usually the opposite. We are using `len()` instead of `.chars().count()` so it sometimes make trouble with non-ascii strings.